fix: исправлен тип данных

+++ Helpers.h
@@ -1,6 +1,6 @@
 #pragma once
 
-int sum(int a, int b)
+float sum(float a, float b)
 {
 	return std::pow(a + b, 2);
 }
